undo serialized extension rename
