+++ hugr-py/src/hugr/ext.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING, Any
 
 from semver import Version
 
-import hugr.serialization.ext as ext_s
+import hugr.serialization.extension as ext_s
 from hugr import tys, val
 from hugr.utils import ser_it
